// src/modules/appointment/types/appoinments-api.ts
import type { DateTime } from "luxon";

// Ideally API types would be autogenerated from the contract specification

export type AppointmentDateRangeGrouped = Record<
  string,
  AppointmentDateRange[]
>;

export interface AppointmentDateRange {
  Start: DateTime;
  End: DateTime;
  Taken: boolean;
}

export interface GetWeeklySlotsRequest {
  from: DateTime;
  to: DateTime;
}

export interface GetWeeklySlotsResponse {
  Start: string;
  End: string;
  Taken: boolean;
}

export interface PostBookSlotRequest {
  Start: string;
  End: string;
  Comments: string;
  Patient: Patient;
}

export interface Patient {
  Name: string;
  SecondName: string;
  Email: string;
  Phone: string;
}